Move `_get_param_values` function to utils module
This utility function can be used regardless of visualization backend,
so it should be available in general util module
